requirements.txt
streamlit
streamlit-lottie
google-generativeai
dotenv